backend/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
python-decouple==3.8
django-cors-headers==4.8.0